Features
* add support for 4G mem
* dummy ext2 support
* add mmap
* fix in exec - args were broken due to using standard strcpy returning
  beginning of dest instead of end of dest
* fix for first visible line in console
* sh looks for binaries in /bin
* fix for overflown lines in console
* fix for overflown early printk buffer

+++ drivers/char/console.h
@@ -4,7 +4,7 @@
 #include <kernel/fs.h>
 #include "console_driver.h"
 
-int console_write(struct file*, const char* buffer, int size);
+int console_write(struct file*, const char* buffer, size_t size);
 void console_putch(uint8_t c);
 int console_init(console_driver_t* driver);
